+++ src/main/java/ru/vyarus/guice/persist/orient/repository/command/async/AsyncQueryMethodExtension.java
@@ -2,7 +2,10 @@ package ru.vyarus.guice.persist.orient.repository.command.async;
 
 import com.google.common.base.Strings;
 import com.orientechnologies.orient.core.command.OCommandRequest;
+import com.orientechnologies.orient.core.command.OCommandResultListener;
 import com.orientechnologies.orient.core.sql.query.OSQLAsynchQuery;
+import com.orientechnologies.orient.core.sql.query.OSQLNonBlockingQuery;
+import ru.vyarus.guice.persist.orient.repository.command.async.listener.mapper.AsyncQueryListener;
 import ru.vyarus.guice.persist.orient.repository.command.core.AbstractCommandExtension;
 import ru.vyarus.guice.persist.orient.repository.command.core.spi.CommandMethodDescriptor;
 import ru.vyarus.guice.persist.orient.repository.command.core.spi.SqlCommandDescriptor;
@@ -13,6 +16,7 @@ import ru.vyarus.guice.persist.orient.repository.core.spi.DescriptorContext;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
+import java.util.concurrent.Future;
 
 import static ru.vyarus.guice.persist.orient.repository.core.MethodDefinitionException.check;
 
@@ -25,6 +29,11 @@ import static ru.vyarus.guice.persist.orient.repository.core.MethodDefinitionExc
 @Singleton
 public class AsyncQueryMethodExtension extends AbstractCommandExtension<CommandMethodDescriptor, AsyncQuery> {
 
+    /**
+     * Extension key used to store query "blocking" setting.
+     */
+    public static final String EXT_BLOCKING = "ASYNC_QUERY_BLOCKING";
+
     @Inject
     public AsyncQueryMethodExtension(final SpiService spiService) {
         super(spiService);
@@ -34,19 +43,40 @@ public class AsyncQueryMethodExtension extends AbstractCommandExtension<CommandM
     public CommandMethodDescriptor createDescriptor(final DescriptorContext context, final AsyncQuery annotation) {
         final CommandMethodDescriptor descriptor = new CommandMethodDescriptor();
         descriptor.command = Strings.emptyToNull(annotation.value());
+        final boolean blocking = annotation.blocking();
+        descriptor.extDescriptors.put(EXT_BLOCKING, blocking);
+        checkReturnType(context.method.getReturnType(), blocking);
 
         analyzeElVars(descriptor, context);
         analyzeParameters(descriptor, context);
 
-        // Listen extension will check that method is void
         check(descriptor.extDescriptors.get(ListenParamExtension.KEY) != null,
-                "Required @%s parameter not defined", Listen.class.getSimpleName());
+                "Required @%s parameter of type %s or %s not defined", Listen.class.getSimpleName(),
+                OCommandResultListener.class.getSimpleName(), AsyncQueryListener.class.getSimpleName());
         return descriptor;
     }
 
     @Override
     protected OCommandRequest createQueryCommand(final CommandMethodDescriptor descriptor,
                                                  final SqlCommandDescriptor desc) {
-        return new OSQLAsynchQuery(desc.command);
+        final boolean blocking = (Boolean) descriptor.extDescriptors.get(EXT_BLOCKING);
+        // correct listener will be set by @Listen extension
+        if (blocking) {
+            return new OSQLAsynchQuery(desc.command);
+        } else {
+            return new OSQLNonBlockingQuery(desc.command, null);
+        }
+    }
+
+    private void checkReturnType(final Class<?> returnType, final boolean blocking) {
+        if (blocking) {
+            check(void.class.equals(returnType) || Void.class.equals(returnType),
+                    "Async query method must be void, because no results returned from query");
+        } else {
+            check(void.class.equals(returnType) || Void.class.equals(returnType)
+                            || Future.class.equals(returnType),
+                    "Non blocking async query method may be only void or return %s because "
+                            + "no results returned from query", Future.class.getSimpleName());
+        }
     }
 }