+++ src/test/groovy/ru/vyarus/guice/persist/orient/repository/command/async/advanced/AdvancedAsyncExecutionTest.groovy
@@ -5,8 +5,6 @@ import com.tinkerpop.blueprints.Vertex
 import ru.vyarus.guice.persist.orient.AbstractTest
 import ru.vyarus.guice.persist.orient.db.transaction.template.SpecificTxAction
 import ru.vyarus.guice.persist.orient.repository.command.async.listener.mapper.AsyncQueryListener
-import ru.vyarus.guice.persist.orient.repository.core.MethodExecutionException
-import ru.vyarus.guice.persist.orient.repository.core.ext.service.result.converter.ResultConversionException
 import ru.vyarus.guice.persist.orient.support.model.Model
 import ru.vyarus.guice.persist.orient.support.model.VertexModel
 import ru.vyarus.guice.persist.orient.support.modules.BootstrapModule
@@ -22,24 +20,28 @@ import javax.inject.Inject
 @UseModules([RepositoryTestModule, BootstrapModule])
 class AdvancedAsyncExecutionTest extends AbstractTest {
 
+    static boolean remote
+
     @Inject
     AdvancedAsyncCases repository
+
+    void setupSpec() {
+        remote = false
+    }
 
     def "Check custom async query listener"() {
 
         when: "calling async query"
         def listener = new CollectingListener<Model>() {}
         repository.select(listener)
-        sleep(70) // required for remote test variation
-        then: "query executed synchronously, async execution possible with remote only"
+        then: "query executed synchronously"
         listener.results.size() == 10
         listener.results.first() instanceof Model
 
         when: "calling async query for objects"
         listener = new CollectingListener<ODocument>() {}
         repository.selectDoc(listener)
-        sleep(70)
-        then: "query executed synchronously, async execution possible with remote only"
+        then: "query executed synchronously"
         listener.results.size() == 10
         listener.results.first() instanceof ODocument
 
@@ -47,8 +49,7 @@ class AdvancedAsyncExecutionTest extends AbstractTest {
         when: "calling async query for strings"
         listener = new CollectingListener<String>() {}
         repository.selectProjection(listener)
-        sleep(70)
-        then: "query executed synchronously, async execution possible with remote only"
+        then: "query executed synchronously"
         listener.results.size() == 10
         listener.results.first() instanceof String
 
@@ -58,7 +59,6 @@ class AdvancedAsyncExecutionTest extends AbstractTest {
         } as SpecificTxAction)
         listener = new CollectingListener<Vertex>() {}
         repository.selectVertex(listener)
-        sleep(70)
         then: "done"
         listener.results.size() == 1
         listener.results.first() instanceof Vertex
@@ -69,7 +69,6 @@ class AdvancedAsyncExecutionTest extends AbstractTest {
         // here listener generic is impossible to resolve directly so parameter declaration will be used instead
         def listener = new CollectingListener<Model>()
         repository.select(listener)
-        sleep(70)
         then: "done"
         listener.results.size() == 10
         listener.results.first() instanceof Model
@@ -81,7 +80,6 @@ class AdvancedAsyncExecutionTest extends AbstractTest {
         // Model, when declared VersionedEntity
         def listener = new CollectingListener<Model>() {}
         repository.selectPolymorphic(listener)
-        sleep(70)
         then: "done"
         listener.results.size() == 10
         listener.results.first() instanceof Model
@@ -91,13 +89,84 @@ class AdvancedAsyncExecutionTest extends AbstractTest {
         // here is impossible to detect correct generic
         listener = new CollectingListener<Model>()
         repository.selectPolymorphic(listener)
-        sleep(70)
-        then: "error"
-        def ex = thrown(MethodExecutionException)
-        while(!(ex instanceof ResultConversionException)) {ex = ex.getCause()}
-        ex.message == 'Failed to convert ODocument to VersionedEntity'
+        then: "error detected but muted"
+        listener.results.isEmpty()
     }
 
+
+    def "Check non blocking case"() {
+
+        when: "call non blocking query with object conversion"
+        def listener = new CollectingListener<Model>()
+        def res = repository.selectNonBlock(listener).get()
+        then: "done"
+        res.size() == 10
+        res.first() instanceof Model
+    }
+
+
+    def "Check non blocking with data modification"() {
+
+        when: "call non blocking query and modify db in listener"
+        def res = repository.selectNonBlock(new AsyncQueryListener<Model>() {
+            @Override
+            boolean onResult(Model result) {
+                result.name = result.name + '_changed'
+                // db modification
+                context.connection.save(result)
+                return true
+            }
+
+            @Override
+            void onEnd() {
+            }
+        }).get()
+        then: "db modifications are not allowed so listener will fail with an exception, but for local all ok"
+        res.size() == (remote ? 0 : 10)
+    }
+
+    def "Check non blocking with extra query"() {
+
+        when: "call non blocking query and use extra query in listener"
+        def res = repository.selectNonBlock(new AsyncQueryListener<Model>() {
+            @Override
+            boolean onResult(Model result) {
+                result.name = result.name + '_changed'
+                // extra db query
+                context.connection.browseClass(Model.class)
+                return true
+            }
+
+            @Override
+            void onEnd() {
+            }
+        }).get()
+        then: "db operations are not allowed for non blocking async connection, but all ok for local conn"
+        res.size() == (remote ? 0 : 10)
+    }
+
+    def "Check blocking with data modification"() {
+
+        when: "call blocking query and modify db in listener"
+        def res = []
+        repository.select(new AsyncQueryListener<Model>() {
+            @Override
+            boolean onResult(Model result) {
+                result.name = result.name + '_changed'
+                // db modification
+                context.connection.save(result)
+                res << result
+                return true
+            }
+
+            @Override
+            void onEnd() {
+            }
+        })
+        then: "done"
+        res.size() == 10
+        res.first() instanceof Model
+    }
 
     // IMPORTANT: if listener would be used directly type will be resolved from parameter
     // always create anonymous class to preserve selected generic